requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.1.0
black==19.10b0
click==7.1.2
Django==3.1
django-cors-headers==3.4.0
django-environ==0.4.5
djangorestframework==3.11.1
flake8==3.8.3
mccabe==0.6.1
pathspec==0.8.0
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
regex==2020.7.14
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
